agents/a2a-business-intelligence/pyproject.toml
[project]
name = "a2a-business-intelligence"
version = "1.0.0"
description = "A2A Business Intelligence Agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "openai>=1.57.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"